Update and rename LICENSE to LICENSE.md
Since this package utilizes PIL to provide its functionality (and is thus basically a kind of front-end to it), I am fairly confident that it counts as derivative work. That is not an issue though as PIL is permissively licensed. However, the one thing that needs to be done is include PIL's license properly, which is what this patch takes care off.