Update torch_transform and tests
There was a bug in the `torch_transform` function where the image was
being re-enclosed in a list multiple times (the respective methods on
the expected image then failed because the items were instead themselves
lists).

The bug was never caught by the tests because only one transform was
ever performed. The tests have also been updated to now catch this.